+++ R/estimate_density.R
@@ -40,6 +40,9 @@
 #' library(rstanarm)
 #' model <- stan_glm(mpg ~ wt + gear, data = mtcars, chains = 2, iter = 200)
 #' head(estimate_density(model))
+#'
+#' library(emmeans)
+#' head(estimate_density(emtrends(model, ~1, "wt")))
 #' \dontrun{
 #' # brms models
 #' # -----------------------------------------------
@@ -142,6 +145,17 @@ estimate_density.data.frame <- function(x, method = "kernel", precision = 2^10, 
   out[, c("Parameter", "x", "y")]
 }
 
+#' @export
+estimate_density.emmGrid <- function(x, method = "kernel", precision = 2^10, extend = FALSE, extend_scale = 0.1, bw = "SJ", ...) {
+  if (!requireNamespace("emmeans")) {
+    stop("Package \"emmeans\" needed for this function to work. Please install it.")
+  }
+  x <- as.data.frame(as.matrix(emmeans::as.mcmc.emmGrid(x, names = FALSE)))
+
+  estimate_density(x, method = method, precision = precision,
+                              extend = extend, extend_scale = extend_scale,
+                              bw = bw, ...)
+}