+++ R/hdi.R
@@ -53,6 +53,9 @@
 #' model <- stan_glm(mpg ~ wt + gear, data = mtcars, chains = 2, iter = 200)
 #' hdi(model)
 #' hdi(model, ci = c(.80, .90, .95))
+#'
+#' library(emmeans)
+#' hdi(emtrends(model, ~1, "wt"))
 #' \dontrun{
 #' library(brms)
 #' model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
@@ -98,6 +101,18 @@ hdi.data.frame <- function(x, ci = .89, verbose = TRUE, ...) {
   dat <- .compute_interval_dataframe(x = x, ci = ci, verbose = verbose, fun = "hdi")
   attr(dat, "object_name") <- deparse(substitute(x), width.cutoff = 500)
   dat
+}
+
+#' @rdname hdi
+#' @export
+hdi.emmGrid <- function(x, ci = .89, verbose = TRUE, ...) {
+  if (!requireNamespace("emmeans")) {
+    stop("Package \"emmeans\" needed for this function to work. Please install it.")
+  }
+  xdf <- as.data.frame(as.matrix(emmeans::as.mcmc.emmGrid(x, names = FALSE)))
+  out <- hdi(xdf , ci = ci, verbose = verbose, ...)
+  attr(out, "object_name") <- deparse(substitute(x), width.cutoff = 500)
+  out
 }