+++ R/p_rope.R
@@ -25,6 +25,9 @@
 #' library(rstanarm)
 #' model <- stan_glm(mpg ~ wt + gear, data = mtcars, chains = 2, iter = 200)
 #' p_rope(model, precision = 1)
+#'
+#' library(emmeans)
+#' p_rope(emtrends(model, ~1, "wt"))
 #' \dontrun{
 #' library(brms)
 #' model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
@@ -115,6 +118,19 @@ p_rope.data.frame <- function(x, range = "default", precision = .1, ...) {
     stringsAsFactors = FALSE
   )
   class(out) <- c("p_rope", class(out))
+  out
+}
+
+#' @rdname p_rope
+#' @export
+p_rope.emmGrid <- function(x, range = "default", precision = .1, ...) {
+  if (!requireNamespace("emmeans")) {
+    stop("Package \"emmeans\" needed for this function to work. Please install it.")
+  }
+  xdf <- as.data.frame(as.matrix(emmeans::as.mcmc.emmGrid(x, names = FALSE)))
+
+  out <- p_rope(xdf , range = range, precision = precision, ...)
+  attr(out, "object_name") <- deparse(substitute(x), width.cutoff = 500)
   out
 }