+++ R/print.bayesfactor_savagedickey.R
@@ -6,6 +6,14 @@ print.bayesfactor_savagedickey <- function(x, digits = 2, log = FALSE, ...) {
 
   if (log) {
     BFE$BF <- log(BFE$BF)
+  }
+
+  xBF <- BFE$BF
+  BFE$BF <- as.character(round(xBF, digits = digits))
+  BFE$BF[abs(xBF) >= 1000 | abs(xBF) < 1 / (10 ^ digits)] <-
+    formatC(xBF, format = "e", digits = digits)[abs(xBF) >= 1000 | abs(xBF) < 1 / (10 ^ digits)]
+
+  if (log) {
     colnames(BFE)[colnames(BFE) == "BF"] <- "log(Bayes Factor)"
   } else {
     colnames(BFE)[colnames(BFE) == "BF"] <- "Bayes Factor"
@@ -15,11 +23,11 @@ print.bayesfactor_savagedickey <- function(x, digits = 2, log = FALSE, ...) {
 
   print.data.frame(BFE, digits = digits, row.names = FALSE)
   cat("---\n")
-  cat(paste0("Evidence Against Test Value: ", round(hypothesis, digits), "\n"))
+  cat("Evidence Against Test Value: ", round(hypothesis, digits), "\n")
   if (direction < 0) {
-    cat(paste0("Left-Sided test\n"))
+    cat("Left-Sided test\n")
   } else if (direction > 0) {
-    cat(paste0("Right-Sided test\n"))
+    cat("Right-Sided test\n")
   }
   invisible(x)
 }